requirements_complete.txt
# AI-Financial-Agent Complete Requirements
# Updated for OCR Implementation Phase 1

# Core FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Database
motor>=3.3.0
pymongo>=4.6.0
mongoengine>=0.27.0

# Authentication & Security
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# M-Pesa Integration 
requests>=2.31.0
httpx>=0.25.0

# AI and Machine Learning
google-generativeai>=0.3.2

# OCR Core Dependencies
opencv-python>=4.8.1
pytesseract>=0.3.10
easyocr>=1.7.0
paddleocr>=2.7.0
Pillow>=10.1.0
pdf2image>=1.16.3
python-magic>=0.4.27

# Image Processing & Computer Vision
scikit-image>=0.22.0
numpy>=1.24.4
supervision>=0.16.0

# Background Processing
redis>=5.0.1
celery[redis]>=5.3.4
flower>=2.0.1

# Async & Utilities
aiofiles>=23.2.0
python-dateutil>=2.8.2
emails>=0.6.0

# Monitoring & Metrics
prometheus-client>=0.19.0

# Communication Services
sendgrid>=6.11.0
africastalking>=1.2.5

# Development & Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0

# Additional OCR Enhancement
# These are optional but recommended for better OCR results
layoutparser>=0.3.4
transformers>=4.35.0
torch>=2.1.0